Make the SubmitButton component, use the theme palette
The `SubmitButton` component was setting the `customButtonTextColor`
(or the `customButtonBackgroundColor`) attribute whenever a colour was
set. This would override any colour classes set to use the theme
palette, like `has-primary-color`. As a result the colours wouldn't
update as the theme was changed.

This also meant that a few components were only rendering the button
using the custom colour attributes, and so would suffer from the same
problem.

Additionally, the Calendly attribute validation was not allowing colour
names like `primary`. For now I've removed the validation as everything
is properly escaped, but in future we should probably create a better
regex to check for hex values or hyphenated lowercase words.